refactor(migrations): apply review feedback
- Let `MIGRATION_STEPS` own the backend-to-visible step mapping, so
  `currentPosition` keeps one special case instead of three.
- Route `StepContent` off `migration.step` alone, so branch order no
  longer carries meaning.
- Fetch the import outcome once and pass it down.
- Use `ButtonGroup` for the two actions, which stacks on narrow screens.
- Reuse `CountEntity` and share `plural` instead of redeclaring both.
- Read the records listing on the read session, like its siblings.
- Split the record-count hooks into their own module, putting both files
  back under the 250-line limit.
